+++ lab_agent/auth/__init__.py
@@ -11,6 +11,9 @@ from .models import CurrentUser, IdentityClaims
 from .service import AuthenticationService
 from .email_otp import EmailDomainPolicy, EmailOtpService, IssuedOtpChallenge, OtpVerificationResult
 from .sessions import IssuedSession, SessionService
+from .email_sender import EmailSender, EmailDeliveryError, EmailAuthenticationError, EmailSenderConfigurationError
+from .smtp_sender import SmtpEmailSender
+from .email_otp_delivery import EmailOtpDeliveryService, EmailDeliveryResult
 
 __all__ = [
     "AuthenticationError",
@@ -27,4 +30,11 @@ __all__ = [
     "OtpVerificationResult",
     "IssuedSession",
     "SessionService",
+    "EmailSender",
+    "EmailDeliveryError",
+    "EmailAuthenticationError",
+    "EmailSenderConfigurationError",
+    "SmtpEmailSender",
+    "EmailOtpDeliveryService",
+    "EmailDeliveryResult",
 ]